moved mocks folder inside mdm
